# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东南大学成贤学院学生成绩录入及学生进行查询系统的标准API接口文档",
        "title": "成绩录入及查询系统标准接口文档",
        "termsOfService": "暂缺",
        "contact": {
            "name": "黄浩轩",
            "url": "https://github.com/huanghaoxuan",
            "email": "huanghaoxuan1998@outlook.com"
        },
        "license": {
          "name": "GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007",
          "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:5201",
    "basePath": "/api",
    "paths": {
...
      "/studentInformation/upload": {
        "post": {
          "description": "上传学生信息表格文件，批量添加学生信息",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "学籍信息"
          ],
          "summary": "上传学生信息表格文件",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "file",
              "description": "文件",
              "name": "file",
              "in": "formData",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/teacherInformation/delete/{id}": {
        "get": {
          "description": "删除一条老师信息",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "老师信息"
          ],
          "summary": "删除一条老师信息",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "主键",
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "/teacherInformation/insert": {
        "post": {
          "description": "增加老师信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "老师信息"
          ],
          "summary": "增加老师信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "教师工号",
              "name": "teacherId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "姓名",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在学院或部门",
              "name": "department",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/teacherInformation/selectByPage": {
        "post": {
          "description": "分页查询老师信息，如果查询第一页，返回总条数，条件非必需",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "老师信息"
          ],
          "summary": "分页查询老师信息",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "查询页码",
              "name": "pageNum",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "每页条数",
              "name": "pageSize",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "教师工号",
              "name": "teacherId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "姓名",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在学院或部门",
              "name": "department",
              "in": "formData"
            }
          ]
        }
      },
      "/teacherInformation/update": {
        "post": {
          "description": "更新老师信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "老师信息"
          ],
          "summary": "更新老师信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "教师工号",
              "name": "teacherId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "姓名",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在学院或部门",
              "name": "department",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
}